chore: 어드민 보조 페이지 App Router 병렬 이관
