+++ apps/admin/src/app/chat-socket/page.tsx
@@ -1,0 +1,10 @@
+import { RequireAdminSession } from "@/components/features/auth/RequireAdminSession";
+import { ChatSocketPageContent } from "@/components/features/chat-socket/ChatSocketPageContent";
+
+export default function ChatSocketPage() {
+	return (
+		<RequireAdminSession>
+			<ChatSocketPageContent />
+		</RequireAdminSession>
+	);
+}
